Create extract_text_from_image.py
File has the code which takes in the snipped image and then copies the table to clipboard which can than be copied to excel. Very simplified version of the code I have. Get this running for now so that later we can just add the files and make it work.